infrastructure/modules/data-mock/python/requirements.txt
functions_framework
google-cloud-storage
google-cloud-pubsub
google-cloud-logging
pandas